+++ example/projection/spring-reactor-transactional-projection-mongodb/src/main/java/org/occurrent/example/eventstore/mongodb/spring/reactor/transactional/TransactionalProjectionsWithSpringAndMongoDBApplication.java
@@ -17,6 +17,7 @@
 package org.occurrent.example.eventstore.mongodb.spring.reactor.transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.annotation.PostConstruct;
 import org.occurrent.eventstore.api.reactor.EventStore;
 import org.occurrent.eventstore.mongodb.spring.reactor.EventStoreConfig;
 import org.occurrent.eventstore.mongodb.spring.reactor.ReactorMongoEventStore;
@@ -29,8 +30,6 @@ import org.springframework.data.mongodb.ReactiveMongoTransactionManager;
 import org.springframework.data.mongodb.core.MongoOperations;
 import org.springframework.data.mongodb.core.ReactiveMongoTemplate;
 import org.springframework.data.mongodb.repository.config.EnableReactiveMongoRepositories;
-
-import javax.annotation.PostConstruct;
 
 /**
  * Bootstrap the application