+++ subscription/mongodb/native/blocking-competing-consumer-strategy/src/main/java/org/occurrent/subscription/mongodb/spring/blocking/NativeMongoLeaseCompetingConsumerStrategy.java
@@ -2,12 +2,12 @@ package org.occurrent.subscription.mongodb.spring.blocking;
 
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
+import jakarta.annotation.PreDestroy;
 import org.bson.BsonDocument;
 import org.occurrent.retry.RetryStrategy;
 import org.occurrent.subscription.api.blocking.CompetingConsumerStrategy;
 import org.occurrent.subscription.mongodb.blocking.ccs.internal.MongoLeaseCompetingConsumerStrategySupport;
 
-import javax.annotation.PreDestroy;
 import java.time.Clock;
 import java.time.Duration;
 import java.util.Objects;